server/requirements.txt
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
binaryornot==0.4.4
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==46.0.1
daphne==4.2.1
Django==5.2.6
django-axes==8.0.0
django-browser-reload==1.20.0
django-cors-headers==4.9.0
django-tailwind==4.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
honcho==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytokens==0.1.10
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
ruff==0.13.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zope.interface==8.0
